Split Out Start Task
+ Moves start task to separate file.
+ Adds requires for start task.
